Fix some warnings

+++ Xenity_Engine/Source/editor/ui/menus/build_settings_menu.cpp
@@ -156,12 +156,12 @@ void BuildSettingsMenu::Draw()
 				cursorPos.y += 5;
 				ImGui::SetCursorPosX(cursorPos.x + 5);
 				ImGui::SetCursorPosY(cursorPos.y);
-				ImGui::Image((ImTextureID)platform.icon->GetTextureId(), ImVec2(imageSize, imageSize), ImVec2(0, 0), ImVec2(1, 1), tint);
+				ImGui::Image((ImTextureID)(size_t)platform.icon->GetTextureId(), ImVec2(imageSize, imageSize), ImVec2(0, 0), ImVec2(1, 1), tint);
 
 				// Draw icon shadow
 				ImGui::SetCursorPosX(cursorPos.x + 5 + 2);
 				ImGui::SetCursorPosY(cursorPos.y + 2);
-				ImGui::Image((ImTextureID)platform.icon->GetTextureId(), ImVec2(imageSize, imageSize), ImVec2(0, 0), ImVec2(1, 1), ImVec4(0, 0, 0, 0.2f));
+				ImGui::Image((ImTextureID)(size_t)platform.icon->GetTextureId(), ImVec2(imageSize, imageSize), ImVec2(0, 0), ImVec2(1, 1), ImVec4(0, 0, 0, 0.2f));
 
 				// Draw text
 				ImGui::SameLine();
